[FIX] Admin's signature added to notification emails when checking 'add user signature' on an email template
When 'add user signature' is enabled on an email template in Odoo 10, the signature
is added to the body as rendered in the compose wizard popup window. When sending out
the email notifications for the message, another signature is added, only this time it is
the administrator's signature as we are in sudo scope.
In Odoo 12, the proper user's signature is always added to the outgoing email notification
and the signature setting on the email template should in fact always be disabled except
on mass mailings (as the comment already denotes at this point).