[ADD] Support big integer fields and in particular bigserial id columns
A serious ERP system needs to support tables with more than 2.1 billion
records.

This change allows the developer to set the model attribute `_bigint_id`
to True to set, or update the sql data type of its `id` column to int8.

This will also recursively update the column type of the columns that have
a foreign key constraint on that `id` column (Many2one and Many2many).

Many2oneReference fields are bigint by default.

While the XML-RPC standard does not officially support integer values that
exceed the signed 32 bits limits, this change patches the Python XML-RPC
client to apply, if necessary, the <i8> extension that is also applied by
the Apache Java implementation of the protocol.