[FIX] core: fix translate override
if the 'translate' attribute of a field in module A is overridden by another
field in module B from `translate=False` to `translate=True`, while upgrading
module A, the info of module B hasn't been loaded.
As a result, the ORM thinks the field is `translate=False` and changes the
column type from jsonb to varchar and drops all translations.

This commit fixes the bug by checking if the `translate` was set to a truthy
value in the database and override the existing attribute

some potential inconvenience
1. if a developer changes `translate=True` to `translate=False`, he doesn't have
   to change the column type, but he should change `ir_model_fields.translate`'s
   column value to `false` if there is no override 'translate=True' from any
   other module
2. after uninstall the override module, the `translate=True` override will be
   remained
3. override `translate=False` to `translate=xml_translate/html_translated`
   may cause unexpected behavior. It is not supported